struct(T5): split actions.rs into studio/live/business/rest
Pure move of Game action methods into actions/{mod,studio,live,business,rest}.rs.
execute_action stays the single match hub in actions/mod.rs. Nested modules
use pub(in crate::game) so turn/tests keep the same call sites. Marks T5 done.
